Bug: 202323961
Merged-In: I154420eb5eb7e14c67f2f5deff02104b481e6ca6

+++ tests/android_car_api_test/src/android/car/apitest/CarUxRestrictionsConfigurationTest.java
@@ -407,7 +407,7 @@ public class CarUxRestrictionsConfigurationTest {
                 + "\"passenger_moving_restrictions\":[{\"req_opt\":true,\"restrictions\":57}],"
                 + "\"passenger_unknown_restrictions\":[{\"req_opt\":true,\"restrictions\":510}]}";
         CarUxRestrictionsConfiguration expectedConfig = new Builder()
-                .setPhysicalPort((byte) 1)
+                .setPhysicalPort(1)
                 .setMaxContentDepth(2)
                 .setMaxCumulativeContentItems(20)
                 .setMaxStringLength(21)
@@ -723,7 +723,7 @@ public class CarUxRestrictionsConfigurationTest {
     @Test
     public void testParcelableConfiguration() {
         CarUxRestrictionsConfiguration config = new CarUxRestrictionsConfiguration.Builder()
-                .setPhysicalPort((byte) 1)
+                .setPhysicalPort(1)
                 .setMaxStringLength(1)
                 .setMaxCumulativeContentItems(1)
                 .setMaxContentDepth(1)
